+++ CheckersClient/src/main/java/com/CheckersGame/Client/View/GameViewComponents/Fields/Field.java
@@ -1,4 +1,4 @@
-package com.CheckersGame.Client.View;
+package com.CheckersGame.Client.View.GameViewComponents.Fields;
 
 import javafx.scene.layout.Background;
 import javafx.scene.layout.BackgroundFill;
